dump-restore scripts moved to commands.
